feat: task.c functions made 64 bit compatible

+++ src/kernel.h
@@ -1,6 +1,8 @@
 #ifndef KERNEL_H
 #define KERNEL_H
 
+#include <stdint.h>
+
 #define VGA_WIDTH 80
 #define VGA_HEIGHT 20
 #define MARROWOS_MAX_PATH 108
@@ -12,8 +14,8 @@ void kernel_page();
 void kernel_registers();
 void terminal_writechar(char c, char color);
 
-#define ERROR(value) (void *)(value)
-#define ERROR_I(value) (int)(value)
-#define ISERR(value) ((int)value < 0)
+#define ERROR(value) ((void *)(intptr_t)(value))
+#define ERROR_I(value) ((int)(intptr_t)(value))
+#define ISERR(value) ((int)(intptr_t)(value) < 0)
 
 #endif
